pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["Alex Suslov <suslov.lexx@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3"
django = "5.1.11"
djangorestframework = "3.15.2"
gunicorn = "23.0.0"
pillow = "^11.1.0"
django-filter = "25.1"
drf-spectacular = "0.28.0"
psycopg = "^3.2.9"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
django-debug-toolbar = "5.1.0"
faker = "^37.3.0"
mypy = "^1.17.1"
django-stubs = "^5.2.2"
djangorestframework-stubs = "^3.16.1"
django-stubs-ext = "^5.2.2"
django-filter-stubs = "^0.1.3"
python-decouple-stubs = "^3.8.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
show_error_context = false
disable_error_code = ["override",]


[tool.django-stubs]
django_settings_module = "mainsite.settings"
